repository int to long fixed

+++ backend/src/main/java/com/example/sellit/repositories/AllegroIntegrationDataRepository.java
@@ -3,5 +3,5 @@ package com.example.sellit.repositories;
 import com.example.sellit.models.AllegroIntegrationData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AllegroIntegrationDataRepository extends JpaRepository<AllegroIntegrationData, Integer> {
+public interface AllegroIntegrationDataRepository extends JpaRepository<AllegroIntegrationData, Long> {
 }
