+++ backend/src/main/java/com/example/sellit/repositories/OpenAIIntegrationDataRepository.java
@@ -4,5 +4,5 @@ import com.example.sellit.models.OpenAIIntegrationData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface OpenAIIntegrationDataRepository extends JpaRepository<OpenAIIntegrationData, Integer>{
+public interface OpenAIIntegrationDataRepository extends JpaRepository<OpenAIIntegrationData, Long>{
 }
