+++ backend/src/main/java/com/example/sellit/repositories/TemplatesRepository.java
@@ -3,5 +3,5 @@ package com.example.sellit.repositories;
 import com.example.sellit.models.Templates;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TemplatesRepository extends JpaRepository<Templates, Integer> {
+public interface TemplatesRepository extends JpaRepository<Templates, Long> {
 }
